+++ Libplanet/Net/Protocols/KademliaProtocol.cs
@@ -6,6 +6,7 @@ using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using Libplanet.Net.Messages;
+using Libplanet.Net.Transports;
 using Serilog;
 using Random = System.Random;